+++ crates/lunas_generator/src/transformers/inputs.rs
@@ -9,7 +9,7 @@ pub fn generate_input_variable_decl(
     for (i, input) in inputs.iter().enumerate() {
         variables.push(VariableNameAndAssignedNumber {
             name: input.variable_name.clone(),
-            assignment: 2u32.pow(i as u32),
+            assignment: 2u128.pow(i as u32),
         });
     }
     let prop_name = inputs
